pyproject.toml
[project]
name = "thinkflix-briefing"
version = "0.1.0"
description = "Open-source automated multimedia briefing generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "imageio-ffmpeg>=0.5.1",
    "kokoro>=0.9.4",
    "numpy>=1.26.0",
    "pillow>=10.4.0",
    "pydantic>=2.8.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "soundfile>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.6.0",
]

[project.scripts]
briefing = "briefing.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["briefing"]
artifacts = ["briefing/prompts/*.md", "briefing/prompts/*.txt"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
